+++ opentelemetry-prometheus/src/lib.rs
@@ -80,6 +80,8 @@ use opentelemetry::{
     metrics::{registry::RegistryMeterProvider, MetricsError, NumberKind},
     Key, Value,
 };
+use std::env;
+use std::num::ParseIntError;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
@@ -92,13 +94,25 @@ const DEFAULT_CACHE_PERIOD: Duration = Duration::from_secs(0);
 
 const EXPORT_KIND_SELECTOR: ExportKindSelector = ExportKindSelector::Cumulative;
 
+/// Default host used by the Prometheus Exporter when env variable not found
+const DEFAULT_EXPORTER_HOST: &str = "0.0.0.0";
+
+/// Default port used by the Prometheus Exporter when env variable not found
+const DEFAULT_EXPORTER_PORT: u16 = 9464;
+
+/// The hostname for the Promtheus Exporter
+const ENV_EXPORTER_HOST: &str = "OTEL_EXPORTER_PROMETHEUS_HOST";
+
+/// The port for the Prometheus Exporter
+const ENV_EXPORTER_PORT: &str = "OTEL_EXPORTER_PROMETHEUS_PORT";
+
 /// Create a new prometheus exporter builder.
 pub fn exporter() -> ExporterBuilder {
     ExporterBuilder::default()
 }
 
 /// Configuration for the prometheus exporter.
-#[derive(Debug, Default)]
+#[derive(Debug)]
 pub struct ExporterBuilder {
     /// The OpenTelemetry `Resource` associated with all Meters
     /// created by the pull controller.
@@ -122,6 +136,45 @@ pub struct ExporterBuilder {
     ///
     /// If not set a new empty `Registry` is created.
     registry: Option<prometheus::Registry>,
+
+    /// The host used by the prometheus exporter
+    ///
+    /// If not set it will be defaulted to all addresses "0.0.0.0"
+    host: Option<String>,
+
+    /// The port used by the prometheus exporter
+    ///
+    /// If not set it will be defaulted to port 9464
+    port: Option<u16>,
+}
+
+impl Default for ExporterBuilder {
+    fn default() -> Self {
+        let port: Option<u16> = match env::var(ENV_EXPORTER_PORT) {
+            Err(_) => None,
+            Ok(p_str) => p_str
+                .parse()
+                .map_err(|err: ParseIntError| {
+                    let err_msg = format!(
+                        "Unable to parse environment variable {}=\"{}\" - {}. Falling back to default port {}. ",
+                        ENV_EXPORTER_PORT, p_str, err, DEFAULT_EXPORTER_PORT
+                    );
+                    global::handle_error(global::Error::Other(err_msg));
+                    err
+                })
+                .ok(),
+        };
+
+        ExporterBuilder {
+            resource: None,
+            cache_period: None,
+            default_histogram_boundaries: None,
+            default_summary_quantiles: None,
+            registry: None,
+            host: env::var(ENV_EXPORTER_HOST).ok().filter(|s| !s.is_empty()),
+            port,
+        }
+    }
 }
 
 impl ExporterBuilder {
@@ -160,6 +213,22 @@ impl ExporterBuilder {
         }
     }
 
+    /// Set the host for the prometheus exporter
+    pub fn with_host(self, host: String) -> Self {
+        ExporterBuilder {
+            host: Some(host),
+            ..self
+        }
+    }
+
+    /// Set the port for the prometheus exporter
+    pub fn with_port(self, port: u16) -> Self {
+        ExporterBuilder {
+            port: Some(port),
+            ..self
+        }
+    }
+
     /// Set the prometheus registry to be used by this exporter
     pub fn with_registry(self, registry: prometheus::Registry) -> Self {
         ExporterBuilder {
@@ -189,12 +258,25 @@ impl ExporterBuilder {
 
         global::set_meter_provider(controller.provider());
 
-        PrometheusExporter::new(
+        let host = self
+            .host
+            .unwrap_or_else(|| DEFAULT_EXPORTER_HOST.to_string());
+        let port = self.port.unwrap_or(DEFAULT_EXPORTER_PORT);
+
+        let controller = Arc::new(Mutex::new(controller));
+        let collector = Collector::with_controller(controller.clone());
+        registry
+            .register(Box::new(collector))
+            .map_err(|e| MetricsError::Other(e.to_string()))?;
+
+        Ok(PrometheusExporter {
             registry,
             controller,
             default_summary_quantiles,
             default_histogram_boundaries,
-        )
+            host,
+            port,
+        })
     }
 
     /// Sets up a complete export pipeline with the recommended setup, using the
@@ -218,15 +300,23 @@ pub struct PrometheusExporter {
     controller: Arc<Mutex<PullController>>,
     default_summary_quantiles: Vec<f64>,
     default_histogram_boundaries: Vec<f64>,
+    host: String,
+    port: u16,
 }
 
 impl PrometheusExporter {
+    #[deprecated(
+        since = "0.9.0",
+        note = "Please use the ExporterBuilder to initialize a PrometheusExporter"
+    )]
     /// Create a new prometheus exporter
     pub fn new(
         registry: prometheus::Registry,
         controller: PullController,
         default_summary_quantiles: Vec<f64>,
         default_histogram_boundaries: Vec<f64>,
+        host: String,
+        port: u16,
     ) -> Result<Self, MetricsError> {
         let controller = Arc::new(Mutex::new(controller));
         let collector = Collector::with_controller(controller.clone());
@@ -239,6 +329,8 @@ impl PrometheusExporter {
             controller,
             default_summary_quantiles,
             default_histogram_boundaries,
+            host,
+            port,
         })
     }
 
@@ -253,6 +345,16 @@ impl PrometheusExporter {
             .lock()
             .map_err(Into::into)
             .map(|locked| locked.provider())
+    }
+
+    /// Get the exporters host for prometheus.
+    pub fn host(&self) -> &str {
+        self.host.as_str()
+    }
+
+    /// Get the exporters port for prometheus.
+    pub fn port(&self) -> u16 {
+        self.port
     }
 }
 
@@ -288,7 +390,7 @@ impl prometheus::core::Collector for Collector {
                 let number_kind = record.descriptor().number_kind();
                 let instrument_kind = record.descriptor().instrument_kind();
 
-                let desc = get_metric_desc(&record);
+                let desc = get_metric_desc(record);
                 let labels = get_metric_labels(record);
 
                 if let Some(hist) = agg.as_any().downcast_ref::<HistogramAggregator>() {
@@ -457,4 +559,38 @@ fn get_metric_desc(record: &Record<'_>) -> PrometheusMetricDesc {
         .cloned()
         .unwrap_or_else(|| desc.name().to_string());
     PrometheusMetricDesc { name, help }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::env;
+
+    use super::*;
+
+    #[test]
+    fn test_exporter_builder_default() {
+        env::remove_var(ENV_EXPORTER_HOST);
+        env::remove_var(ENV_EXPORTER_PORT);
+        let exporter = ExporterBuilder::default().init();
+        assert_eq!(exporter.host(), "0.0.0.0");
+        assert_eq!(exporter.port(), 9464);
+
+        env::set_var(ENV_EXPORTER_HOST, "prometheus-test");
+        env::set_var(ENV_EXPORTER_PORT, "9000");
+        let exporter = ExporterBuilder::default().init();
+        assert_eq!(exporter.host(), "prometheus-test");
+        assert_eq!(exporter.port(), 9000);
+
+        env::set_var(ENV_EXPORTER_HOST, "");
+        env::set_var(ENV_EXPORTER_PORT, "");
+        let exporter = ExporterBuilder::default().init();
+        assert_eq!(exporter.host(), "0.0.0.0");
+        assert_eq!(exporter.port(), 9464);
+
+        env::set_var(ENV_EXPORTER_HOST, "");
+        env::set_var(ENV_EXPORTER_PORT, "not_a_number");
+        let exporter = ExporterBuilder::default().init();
+        assert_eq!(exporter.host(), "0.0.0.0");
+        assert_eq!(exporter.port(), 9464);
+    }
 }